msm: camera: isp: Fix the return value for CDM timeout error
In case CDM submit times out, the return value needs to be error
code. Without this change, the error is not propagated and the
session continues then causing a wait on requests, adding to the
delay of when the error is caught.

Change-Id: I670c9916e419876d19e0559645569e0837f9d0ce
Signed-off-by: Harsh Shah <harshs@codeaurora.org>
(cherry picked from commit 111567509bc4739017f0dac85c1be21f97e200ff)