drm: msm: sde: Boost cpubw devfreq device when a new frame is ready
The cpubw devfreq device delegates the speed of the DDR bus, so its
speed is directly correlated with frame drops. Boost it when a new frame
is ready to be rendered to the display to reduce overall frame drops.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
(cherry picked from commit 71d9674923e64ecd1699ce7e57e7befe08ae7bd5)