input: touch: ftm4: Fix potential strncat buffer overflow
This strncat takes two buffers with different lengths, and limits itself
to writing the number of bytes in the source buffer (which is guaranteed
to be null-terminated) instead of the number of bytes in the
potentially-smaller destination buffer.

Fix this by limiting this strncat to the number of bytes in the destination
buffer. The source buffer doesn't need to be bothered with a safety check
here since it's guaranteed to be null-terminated.

Change-Id: Ie17ee08bfc6aa99fb0a42cdab168eef533f98aa1
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
(cherry picked from commit c0812ed9c45e06d962ec3b90a96e7fea3fb1fd7b)
(cherry picked from commit fd5b76d5f2b00ea9ff0c82a6bb2debd9132d6bc1)